Changed macOS version + Added fail-fast property to matrix strategy
The macOS job was failing as the docker setup action is not suited for macOS ARM64 based images.
https://github.com/douglascamata/setup-docker-macos-action?tab=readme-ov-file#arm64-processors-m1-m2-m3-series-used-on-macos-14-images-are-unsupported

Hence, now using "macos-latest-large" (macOS 14 as of this commit) image as per the official runner images here:
https://github.com/actions/runner-images/tree/main

-----------

Also, this failure led to that specific job failing and all other jobs being cancelled including the job for ubuntu OS due to the "fail-fast" property being set to true as default.
Hence to allow other jobs to go through, setting the fail-fast property to false:
- https://github.com/orgs/community/discussions/27192#discussioncomment-3254964
- https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idstrategyfail-fast

Signed-off-by: Mahadik, Mukul Chandrakant <MukulChandrakant.Mahadik@nrel.gov>