Increasing timeout to 60 minutes
MacOS docker setup passed finally.

Increasing the timeout since the Downloading and Extracting steps themselves have taken about 20 minutes in the macOS runner.
But right now even for the demo-automated-tests.sh on MacOS its taking close to 30 minutes just for the download + extraction to complete.

As seen in the commit history, when Shankari changed the timeout to 30 minutes, it was because the demo-iso15118-2-ac script itself was taking close to 30 minutes.
https://github.com/EVerest/everest-demo/pull/23/commits/05528d4e76c1dc69f764bfb83459c88624613d3c

Signed-off-by: Mahadik, Mukul Chandrakant <MukulChandrakant.Mahadik@nrel.gov>