Add another job to run automated test scripts
Deleted e2etest.yaml.

Using cicd.yaml as the only workflow file to run another job for automated tests

The new job will run in a new VM instance so first checking if there are any images available already from the previous jobs since "cache-to" is used in the docker build and push step in the matrix jobs.

Signed-off-by: Mahadik, Mukul Chandrakant <MukulChandrakant.Mahadik@nrel.gov>