Added no-cache as true for first build job
I changed the run-test.sh to echo some test value but in the run tests job, the run-test.sh still executes the pytest command from the older version of run-test.sh and does not echo my test value.

Looks like it is using cached version.
This could be because the first build job itself is using cached version.