Commenting out build, pull_policy attributes to see if docker compose builds the image locally or pulls
The reason I had added these in the first place was to forcibly ensure that the image is built locally for testing.
We do not want the automated tests to be run on the latest tagged image in the GHCR repo as this image will not be latest built image using the latest code changes that we actually want to test.

So demonstrating that without these attributes, the image is pulled and not built locally, so running tests on pulled image is not suitable.