See if docker compose builds image locally if there are changes to Dockerfile
For the purposes of this experiment, I've added a test-exit-code.sh script file that I've been using for testing success and failure test cases. I've commented out the code in run-test.sh to just perform a trivial task like printing out the contents of the test-exit-code.sh file.
From my previous runs when I was playing around with getting the docker compose to build the image locally, I saw that without specifying the build and pull_policy attibutes, the image is always pulled from the specific image source - in our case the GHCR registry.

So, the changes I've introduced in this commit, should count as changes to the image being built (manager service image in this case).
And I'm expecting that the workflow run will not build images locally in the docker compose step in the run automated tests job.