Check if docker compose rebuilds image
Using another trivial script as an entrypoint to see if docker compose is able to execute it.
Based on previous runs, it should not be able to even detect the presence of the file since it will not rebuild image locally and instead uses latest image available in GHCR.
So it still uses run-test.sh.

Previous workflow run: https://github.com/MukuFlash03/everest-demo/actions/runs/9912983164/job/27388979064

This also explains why in the previous workflow run, despite commenting out the pytests in run-test.sh, it still executed the tests. This happened since it did not detect changes from the updated run-test.sh as the image was not built locally but instead pulled the GHCR image.