Testing to see if image will be built or pulled in case no changes to image context or Dockerfile
This will help clarify whether we need pull_policy attribute as well.

I see two scenarios:
- Changes to dockerfile, files used to build up image, or in short changes to image.
- Changes to other repo files, like demo scripts, should be alright to not rebuild image.

Now on pushing this commit, we'll be testing scenario 2 since a previous commit actually did modify the Dockerfile and that was Scenario 1.
But right now, no changes to image and hence despite specifying build, image should not be built.

Previous commit:
https://github.com/MukuFlash03/everest-demo/commit/ad8863e0af6dc13996c929fb3911b54e4ac1b84f

But wait, this Scenario 2 only applies to cached images but the docker compose commands run in an isolated container and there would be no reference to the previous images built locally. Hence, we are forcing docker compose to build images always.

So not specifying pull_policy, will let it take the default value of 'missing', which has the following notes from the documentation:
Compose pulls the image only if it's not available in the platform cache.

pull_policy documentation:
https://docs.docker.com/compose/compose-file/05-services/#pull_policy